deleted DuckySprite, health/points/timer moved to Player class, all other content moved to AnimatedSprite class
